Drop unneeded now `withVoids()` plugin
And move its tests to Editor.test.tsx to catch regressions
