+++ packages/slate-editor/src/modules/editor/plugins/index.ts
@@ -4,4 +4,3 @@ export { withNonEmptyValue } from './withNonEmptyValue';
 export type { RichBlocksAwareEditor } from './withRichBlocks';
 export { withRichBlocks } from './withRichBlocks';
 export { withRootElements } from './withRootElements';
-export { withVoids } from './withVoids';
